turn on -Werror=type-limits in our Bazel CPU build
Summary:
We also fix any existing issues. Note that we only do this for the CPU
build because nvcc is considered a C++ toolchain but it does not have
the same flag support. Adding flags to the GPU build will cause nvcc
errors.

Test Plan: Built locally, rely on CI to confirm.

Reviewers: malfet

Subscribers:

Tasks:

Tags:

ghstack-source-id: 741338a82aa2c04bbf6f7880cdf6de0a26819640
Pull Request resolved: https://github.com/pytorch/pytorch/pull/79139