turn on -Werror=missing-field-initializers in our Bazel CPU build
Summary:
We also fix any existing issues. Note that we only do this for the CPU
build because nvcc is considered a C++ toolchain but it does not have
the same flag support. Adding flags to the GPU build will cause nvcc
errors.

Test Plan: Built locally, rely on CI to confirm.

Reviewers: malfet

Subscribers:

Tasks:

Tags:

ghstack-source-id: c81f9d4f8294a763e528410032458ae40261b748
Pull Request resolved: https://github.com/pytorch/pytorch/pull/79155